requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
# PyMuPDF==1.23.8  # Commented out due to Windows build issues

# OpenAI integration
openai==1.3.0

# Basic utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Optional: For advanced features
# spacy==3.7.2
# sentence-transformers==2.2.2
# scikit-learn==1.3.2
# numpy==1.26.2
# pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1


